mon module merlin_test_utils: Lower check frequency timer
It's necessary that we get all checks run properly for us to be
able to make sure that check distribution works properly. In
order for us not to have to wait a full 15 minutes for all tests
to run, we schedule things quite a bit tighter. We can, if we're
so inclined, lower the number of hosts and services per poller
group if the load turns out to be too much to bear.

Signed-off-by: Andreas Ericsson <ae@op5.se>