merlin_apps_utils: Add shorthands for 'bright' colors
We also alias the obviously retardedly named colors to something
niftier, so that 'yellow' becomes an alias for 'bright_brown'
and 'grey' becomes synonymous with 'bright_grey' (as the former
is actually black).

For testability, we also add the method print_colors() which
prints all colors with their designated names.

Signed-off-by: Andreas Ericsson <ae@op5.se>