Always calculate peer id stuff
The peer id stuff might not be strictly required for single-node
systems, but it won't hurt to keep program_status free of any
misinformation.

Easy way to see what changed: mon check distribution used to always
return CRITICAL for single-node systems, while now it will always (as
soon as the CTRL_STALL part is over) return OK. This actually
corresponds to the reality.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>