module: Make the reaper_thread variable module-global
It's nifty to have it as such since it lets random calls in the module
know if they originated from within the reaper thread or from within
the core thread. That way we can know if certain events were triggered
by packets received over the network or if they came about as a result
of something that happened on "our" node.

Signed-off-by: Andreas Ericsson <ae@op5.se>