module: Forward acknowledgement commands again
Since we blocked acknowledgement commands it appears we also added a
bug that caused notifications from being spread at all. This happens
because host and service status events are only ever sent from the
node responsible for handling the checks for that host (or service),
and are otherwise blocked. We handle it quite simply by forwarding
the acknowledgement commands as-is once again.

This fixes issue #4581, but with fallout which will be handled later.

Signed-off-by: Andreas Ericsson <ae@op5.se>