module: Block acknowledgement notifications from the merlin network
Since we're forwarding acknowledgement commands we need to block the
notifications those would generate when we receive one from the
network. Since they will be sent from the command-receiving node
anyway, we're not losing notifications anywhere and we can make sure
the event and status updates get to the receiving end as quickly as
possible.

Signed-off-by: Andreas Ericsson <ae@op5.se>