Add various magic packet destination flags
We want to distribute servicegroup events only to peers and pollers,
and the module should have more or less supreme command of where its
events go. This patch lays the groundwork for such control by adding
various magic flags for the different destinations that we might
want to use. They're not that many, so we sprinkle shorthands for
all of them.

Signed-off-by: Andreas Ericsson <ae@op5.se>