app/conf: Handle b0rked locales
Apparently, at least some OS X machines consider their language and
region to be "UTF-8". For pretty obvious reasons, a sane system cannot
find that region in it's locale database, and will bail out loudly,
which causes the mon command to crash as soon as you run it. Unexpected.

Fix by manually trying to set LC_CTYPES instead of relying on
getpreferredencoding to do it for us, and fall back to locale C.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>