Rework commits 890e53 and 956b64
It appears that the i686 architecture requires that thread-storage
variables are stashed on the stack, or the pthread.h is included
after a lot of other headers, or that the variable stashing the
thread-local storage is never accessed by a longjmp.

Whatever it was, making reaper_thread a global variable and
accessing it (read-only, for cryin' out loud!) caused the
hookinfo struct to be completely messed up in the non-reaper
thread, resulting in instantaneous (read-access) segfaults on
32-bit architectures that happened to receive a string-containing
event from a different node in the merlin network.

With this patch though, it's been happily chugging along for the
past 30 minutes and I can (finally) go to bed.

Signed-off-by: Andreas Ericsson <ae@op5.se>