daemon: Fix stupid typo in find_node()
Since we introduced the "lookup-by-ip-and-portnumber" thing we've
failed horribly to fall back to "lookup-by-ip-only" due to a really
stupid typo where we instead of assigning the node found to the
variable keeping the first found with right ip and instead done it
the other way around, causing us to never ever find it properly.

This patch fixes it so we once again fall back to matching on ip
address.

Signed-off-by: Andreas Ericsson <ae@op5.se>