module: Fix ipc_action_handler
Quite a long way back we changed when the node->state flag is set
relative to the call to the action handler associated with the node.

This particular handler has been using the old way this whole time,
causing spurious and shortlived issues in check distribution since
nodes in busy networks now sometimes never send their own &self to
its peers.

I wouldn't be surprised if this happens to be behind the bug that
causes some checks to never execute. It could easily happen if the
network is busy enough to send events more than every 15 seconds.

Signed-off-by: Andreas Ericsson <ae@op5.se>