module: Stop stalling when daemon disconnects
If the daemon disconnects while we're stalling and before we've
received a CTRL_RESUME packet we can't be sure the daemon actually
got the packet we sent to it, or that it did anything with it. In
order to play it safe, we must in those cases immediately stop stalling
and resend our paths when the daemon comes back online so that it can
run the import on its own.

Signed-off-by: Andreas Ericsson <ae@op5.se>