mon test: Add 'intermission()' method to fake_mesh class
When testing commands in a distributed environment it's important to
allow a little time for the commands to spread to all nodes in the
network. Using a simple sleep helper makes it easy to always allow
the same amount of time for all pauses, depending on the size of the
network.

When sitting and staring at a computer screen it's important to have
something fun to watch. That's why the sleep helper counts down very
quickly and has a little spinning thingie that can occupy ones
attention for almost a full 3.6 seconds. The real reason such things
get implemented is ofcourse that the developer is travelling and
managed to finish his book much faster than anticipated. Oh well.

We do however set a maximum of 30 seconds for intermission sleeping,
since one could successfully argue that if command propagation takes
more than 30 seconds, the test is a failure anyway, almost no matter
how large the environment is.

Signed-off-by: Andreas Ericsson <ae@op5.se>