mon test dist: Add test to make sure imports work properly
We really only test to make sure hosts and services get imported, but
that's enough since we really only want to make sure the import program
gets run at all. The import program in itself has other tests that makes
sure it works flawlessly.

It's a fairly important test for the 'mon test dist' command though,
since we'll want to exit early if the import fails for some reason.

Signed-off-by: Andreas Ericsson <ae@op5.se>