mon test dist: make sure the various test-groups return fals on failure
Previously this was a bit skittish due to the tap class acting weird,
but now that that's fixed we need to make sure we send a proper return
value back. Since the sleep time has gone up quite a lot, the testsuite
now takes a few minutes to run. Valuable time to keep running stuff
that you know won't work properly because the setup failed.

Signed-off-by: Andreas Ericsson <ae@op5.se>