mon test dist: Add and use the --batch option
This can be added to make mon test dist exit without user input
after it's done all the tests. If the tests passed, we clean up
after us. If they didn't, we should probably leave logfiles,
configuration and whatnot around as well, but this is as good a
place as any to start at.

Signed-off-by: Andreas Ericsson <ae@op5.se>