mon log: Make 'mon log import' print something when import fails
Previously we got a very unhelpful SubcommandException traceback.
With this patch, we should at least see the error output of the
failing program, as well as the command we tried to run and a
message stating that it was the other program that failed to
complete its designated actions properly.

Signed-off-by: Andreas Ericsson <ae@op5.se>