ocimp: Handle more complex timeperiods gracefully
Given a timperiod exception like the following
    monday - friday 08:00-17:00
ocimp would believe that this referred to monday, at "- friday
08:00-17:00" o'clock. This is obviously bogus.

However, there was code for fixing timeperiods, it just wasn't being
executed until the second time it spotted a time declaration like the
above, or whenever the exception didn't start with a weekday (which
might be the most common - but not only! - case).

Fix by always re-splitting the timeperiod properly, before checking if
it's an exception or not. Slightly slower, but actually correct.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>