ocimp: Create list of attributes to ignore from objects.cache
Many attributes are duplicated between objects.cache and status.log,
except status.log is the current state. However, we parse objects.cache
after status.log, which caused the current values to be overwritten by
old ones. This means ninja will show the wrong information for one check
interval after a restart.

Fix by painfully generating a list of values that are present in
status.log, that must therefore be ignored when parsing objects.cache.
As this list is only used the second time we traverse an object file,
any bugs in it could maximum cause a column to get the default value,
which is no worse than not doing this in the first place, so this is
pretty harmless.

This fixes bug #5114.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>