nrpe-merlin: Remove asmonitor
It prints a lot of extra output which doesn't sit well with nrpe
and the plugin it's about to execute.

This means we'll have to keep running nrpe as monitor in most
environments, but since that's already implemented in the spec
file no further changes are required to get the merlin checks to
work properly.

Signed-off-by: Andreas Ericsson <ae@op5.se>