mon oconf sync: Assist synergy with config synchronization
This is sort of half-stupid, since synergy rules will only ever
be synced when "mon oconf push" is run, and it will almost
certainly do Bad Things(tm) if several nodes are used to configure
things in the same restart interval. Given that we have three
days to come up with something workable though, this is as good
as it's gonna get before the sync daemon is implemented and
written.

Signed-off-by: Andreas Ericsson <ae@op5.se>