report-data-indexes: Add new combined index, remove redundant
We don't need an index for host_name when we have one for (host_name,
service_description), as the latter covers the former. Add new one for
(event_type, host_name, service_description) - will be big, but the
number of event_types are pretty low, so shouldn't be *that* many times
bigger than (host_name, service_description), so we should get away with
it.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>