node.h: Increase the maximum packet we're able to handle
It doesn't really make any sense to leave this limit at 32KiB
now that we know the protocol as well as the iocache code is
stable and working, so we increase it to half the size of what
we'll (try to) set the send and receive buffers to. Since many
systems seem to have a hard limit of 224KiB for socket buffers
and we attempt to set them to 256, it wouldn't be good to use
more than all of it for a single packet, so we stick with 128.

This "fixes" issue #4713 in the simplest and least flexible way,
which doesn't force us to add dynamic buffer handling and all
its complexities and allocate/deallocate performance downsides.

Signed-off-by: Andreas Ericsson <ae@op5.se>