mon start: Handle synergy config clash when we have masters
In case we have masters and synergy is installed on the system, we
need to remove the synergy configuration from the poller, since it
will be sent to us from the master.

We handle it here by simply removing the cfg_dir line in nagios.cfg
in case it exists. This should be simple enough that it doesn't
mess things up for any opensource users.

Signed-off-by: Andreas Ericsson <ae@op5.se>