module: Keep track of which objects have notified for what
When a notification happens due to a checkresult caused by a check
launched from a peer or poller and the check later fails over to a
different node, recovery notifications sometimes went missing. This
patch at least lets all nodes in the cluster know when some node
sends out a notification. This should cause recovery notifications
to happen no matter if the check fails over to some other node in the
cluster.

The downside is that we're altering the size of the monitored object
state in the process, so the protocol version has to be bumped so we
don't fail (hard) when decoding inbound packets.

Signed-off-by: Andreas Ericsson <ae@op5.se>