compat detection: Send size of our in-core structures
When we start talking to a new endpoint in the communication chain,
we need to tell it what size our monitored_object_state is, so the
node can grok our host and service events, or at least know if it
can't. We also add some few helpers for nodeinfo version and protocol
version so we can be helpful in telling people which verson they
need to upgrade to (at least on the already-upgraded side).

Signed-off-by: Andreas Ericsson <ae@op5.se>