module: Fix state history transfers
Some users have reported that flapping notifications are sometimes
sent out even when a host or service has been stable for a very long
time. Presumably, this happens because two peers have a different
notion of the object's state history, which previously wasn't
transferred between peers or between pollers and masters.

With this patch we hopefully fix issues #4917, #4521 and #4642.

Signed-off-by: Andreas Ericsson <ae@op5.se>