showlog: Make --hide-all --show-notification work properly
Previously the show-notification event didn't set EVT_HOST or
EVT_SERVICE in the event_filter flags, causing all such events
to go missing from the output. This patch fixes it, although we
should probably allow users to hide or show events for only hosts
or only services.

That's a later issue though, and this will work just fine for now.

Signed-off-by: Andreas Ericsson <ae@op5.se>