mon test: Add 'mark' and 'check-mark' commands
The mark command inserts a timestamp, the parent_pid and some random
key/value pairs into any sql table. The check-mark command is meant
to check for such marks and compare their age against a threshold,
but it's not yet implemented.

The idea is to make automated testing of running Nagios instances in
our lab environment a whole lot easier.

Signed-off-by: Andreas Ericsson <ae@op5.se>