logutils: Handle repeated events somewhat gracefully
The previous way was quite retarded when used on experimental code
since such normally logs tons and tons of messages, which we can't
cope with sensibly, and keeping all those messages in memory would
easily cause us to swap before we're done. With this patch we still
don't recognize sequences of messages, but we do handle similarly
repeated messages with somewhat more grace.

Messages are considered similar when the first 20 chars (not counting
the timestamp) are identical, which seems to be a good cutoff to
start with.

This makes imports run in less than a minute again when they used to
take tens of minutes on memory-limited test servers.

Signed-off-by: Andreas Ericsson <ae@op5.se>