daemon: Handle fixed source-port calculations in daemon only
That means moving them from node.c to daemon.c, so the module
doesn't have to worry about them.

And while we're at it, we fix the off-by-one error, the typo and
see to it that loopback interfaces also are covered in this, so
mesh tests still work.

We also add a check so we can warn people when they're being
patently stupid and are adding multiple systems with the same
IP *and* the same port.

Signed-off-by: Andreas Ericsson <ae@op5.se>