daemon: Make connection detection more robust
Previously we used to ignore calling getsockopt() to find out the
link level error in case we failed to get anything sensible out of
getpeername(). That was quite silly, since getpeername() only returns
a good value when we're connected and can only fail due to either a
connection not being completed (which we ignored or forced a disconnect
from), or when the programmer has done something seriously retarded,
like not passing a proper filedescriptor to it, in some other way
fouled up the arguments or when the system is about to commit suicide
due to lack of resources.

We also faced a rather weird error where the socket had some urgent
data for us to read, which we ignored. Ignoring unread urgent data
can cause the socket to linger in kernel space until some timeout
times out (which is what they do, eventually). This patch corrects
that, potentially fixing issue #5456.

As a nice sideeffect, we now always get the real error why we failed
to connect (assuming there is one and it's not just a firewall issue,
which will almost always cause the connection attempt to time out)
and can report that back to whoever sits down and reads the logfiles.

Signed-off-by: Andreas Ericsson <ae@op5.se>