mon test: Let each intermission() caller decide how long to sleep
It makes no sense to sleep forever when we're doing something that
really should be completed in practically notime, so let's avoid
that if at all possible. The previously optional "msg" argument is
now required though, but since we consider it good manners to say
why we're suddenly halting the program, that's actually a good thing.

Signed-off-by: Andreas Ericsson <ae@op5.se>