module: Use Nagios' native calls for locating and deleting downtime
More precisely, we need to locate the entry with the same hostname,
service description, start time and comment. Previously, we used to
do that with a hand-hacked version that traversed the list and did
a ton of comparisons. With this patch, we instead fall back to using
Nagios' internal call that does the same thing.

As a result, the scheduled_downtime could now be made opaque without
us having to bother about it, which is definitely a good thing.

Signed-off-by: Andreas Ericsson <ae@op5.se>