mon oconf nodesplit: Avoid creating identical files
Calculating the content of the files is the (really) heavy part,
whereas we'll know pretty early if multiple pollers share hostgroups,
in which case we can just hardlink the first created file to the new
one in order to avoid doing a lot of hard work that isn't, strictly
speaking, necessary.

Large installations should see a significant improvement in the time
it takes for peered pollers to get their configuration.

Signed-off-by: Andreas Ericsson <ae@op5.se>