mon test: Make config creation a *lot* faster
Previously we used to concatenate a growing buffer with the new
objects continuously, causing object creation to slow down
significantly as the number of created objects grew larger.

With this patch, we create one object at a time and add it to
a list object instead, which we later concatenate once. This
also has a slowdown as N grows larger, but performs a acceptably
for reasonably large N's (5000 or so), where the previous code
would choke around 1500.

Signed-off-by: Andreas Ericsson <ae@op5.se>