sql: Make all sorts of names longer
We've received requests to make some object name fields longer. Make all
object names consistently varchar(255), and hope that's enough for a
while. It's stupid to have long object names, as it reduces performance,
but that's not my problem.

This also reduces some of the field max size confusion we had before:
the host_name field had, in different tables, a max length of 70, 75,
160, 200, and 255 characters.

This *does* change persistent tables, but because we only increase the
column size, it should be pretty much both backwards and forwards
compatible.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>