Kill reaper thread, use nagios' iobroker instead
By hooking into nagios' mainloop through its main iobroker, we no longer
have to worry about threading, as the module will now always be invoked
by nagios when it's needed.

This changes the definition of "stall" somewhat. It used to mean "make
nagios sleep so nothing happens", but that's not really feasible when
all the action goes on in separate threads - instead of preventing us
from missing any status updates, it might cause nagios to miss out on
check results if the communication socket to the workers overflow.

Now, instead, we let nagios work in the background while the merlin
daemon does it's initial import, and have the module binlog anything
that happens while that's going on - this means the daemon must be
finished before 110MB of events have occured, but if it's not, I suspect
we might have other problems.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>