daemonize.c: Do unlink pidfil on shutdown signal
Because both daemon.c and daemonize.c tried to register a handler for
SIGINT/SIGTERM, only one of them was successful, which in this case was
daemon.c. The only thing done in daemonize.c's signal handler was to
remove the pid file, which thus wasn't done, causing it to be left
behind, which isn't pretty. This meant op5kad always thought merlin had
crashed, instantly restarting it, which was never the intended behavior.

Make daemon.c call daemonize.c's previous signal handler, and thus do a
clean shutdown.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>