mon test dist: Make config generation blazing fast
The ridiculous linear slowdown we had earlier was a result of us
doing a linear search in a list to see if the object was already
added. Since the only use the list was intended to be put to was
to get the count of items in it, we might as well save ourselves
the trouble and count the list immediately. We ensure uniqueness
by adding a dictionary where we do lookups.

Now, generating a config with 20k hosts is done in less than two
seconds, which is acceptable, instead of just under five minutes,
which is not.

Signed-off-by: Andreas Ericsson <ae@op5.se>