module: Discard check results that are older than what we have
Accepting them will light up the entire UI with problems if a
poller (or peer) has become isolated for anything but a microscopic
period of time and then comes back online, it will send us a whole
slew of out-of-date check results which we really can't use.

This patch does absolutely nothing if "takeover = no" has been set
for the isolated node, since we in that case won't be getting any
new check results anyway.

Signed-off-by: Andreas Ericsson <ae@op5.se>