ocimp: Use binary lookups when looking for keys
It's a few seconds faster on really large configs (although we're
already spending most of the time running queries, so it won't be a
huge impact). It does make it really easy for us to print the keys
in sorted order, which makes it easy to compare them against keys
obtained from Nagios' header files and see which ones we're missing.

Signed-off-by: Andreas Ericsson <ae@op5.se>