install-merlin.sh: Don't break if upgrade step is too large
A while back, we dropped the id column on report_data. Some time later,
we added a report_data_extras table, the definition of which is a copy
of the report_data table.

If you upgrade from before the id column was dropped to after
report_data_extras was created, report_data_extras would get the id
column, and then it's content couldn't be easily injected into
report_data any longer. Stupid. Stop doing that.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>