new cache file: timeperiods.cache
This is a smaller version of objects.cache, that only contains
timeperiods - these objects are few, self-contained, and aren't
available from livestatus, so hack them in here.

Because this is an Evil Hack, it's is implemented secretly, so we'll be
able to revert this patch ASAP, without anyone noticing. Schhh!!

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>