binlog: stop fsync()ing file-stashed binlog entries
fsync() means we wait for the disk to accept our data before we
move on, which kills performance completely. It also means we
don't get to use the kernel's diskbuffer when we read it back
in to memory, so it's a huge loss in many ways to force data
to hit the disk before we move on.

Signed-off-by: Andreas Ericsson <ae@op5.se>