module: Warn about sub-zero latency
This patch adds a warn_flags field to the merlin_node structure
and turns latency into a signed field once again, which provide
us with a few neat things. First off we can get a log entry of
the fact that there's a clock skew between us and the other node.
Secondly we can mark it as a registered warning and provide that
warning through the query handler, and thirdly we can get the
'mon node status' output to make a bit more sense than showing
some 75 years worth of latency.

Signed-off-by: Andreas Ericsson <ae@op5.se>