merlin_conf: Only force user when one was provided
In particular, when running as "monitor", we'd rather SSH as monitor
than as root, because even though mon sshkey push syncs the monitor
users SSH keys with each other, it doesn't (and shouldn't) add one
machine's monitor user's SSH key to the other machine's root user,
because that would be extra insecure.

So, this finishes up the work started in 4cb35b5, and should resolve
bug #5310 for real.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>