module: Make insane debug logging of inbound events possible
Sometimes (once in a blue moon, or thereabouts), two peered nodes
seem to get disconnected and will refuse any and all attempts to
reconnect and become friends again. All the while, it appears as
if check results really do come in from the other node. With this
patch, we can quite easily set the module to log like crazy and
tell us if it in fact is getting events from its remote peer.

Signed-off-by: Andreas Ericsson <ae@op5.se>