node.c: Only close and set socket to -1 after action handler is run
The module's ipc_action_handler() is responsible for removing the
ipc socket from Nagios' iobroker set, so we must preserve it until
that one's called.

This fixes two pretty major issues; First of them is the fact that
we can now use the query handler even when the merlin daemon isn't
running. The second is that Merlin will now self-heal as intended
even when it disconnects from the daemon for some reason.

Signed-off-by: Andreas Ericsson <ae@op5.se>