mon qh: Make query and get capable building queries
Having to say:

   mon qh query '#wproc wpstats'

will be difficult to explain to the not-so-command-line-savvy crowd
of users. With this patch, one can substitute the above for:

   mon qh query wproc wpstats

and the 'mon qh' app will automagically create a query from it.
Each argument is separated by a single space, so any fancier
whitespace fiddling than that will have to be quoted for the sake
of the shell.

Signed-off-by: Andreas Ericsson <ae@op5.se>