apps/oconf: Make merlin not break when using pollers and changing config
We didn't handle all the cases, so we could end up in a case where we
would neither regenerate the config, nor have cached config.

There was also a force that was misplaced, so we couldn't use that as a
work-around for the above bug.

Also cast to int, because that's enough precision for our uses, and use
exactly equal, because we if the poller cache timestamp is greater than
our most recent change, something is off.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>