Make binlog directory configurable
We used to pretend we grok'ed ipc_binlog_dir, but we really didn't.
Instead we just ignored it when we created binlogs. This patch
adds proper support for configuring it, by setting 'binlog_dir' in
either of the global, module or daemon namespace in merlin.conf,
depending on if one wants to use a common binlog directory or a
separate one for module and daemon.

Since this was always intended to be configurable, I consider this
a bugfix.

Requested-by: Mark Frost <mark.frost1@pepsico.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>