mon: Fix running uncategorized commands
Because we skip over the first arg to 'mon' now, the help processing
'if' needs to look for len < 1 or for 'help' or '--help' at array
index 0 instead of 1. Without this, we always launch into the help
function when trying to run uncategorized commands with no arguments,
such as "mon start".

Signed-off-by: Mark Frost <mark.frost1@pepsico.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>