mon: Add our module dir to sys.path() before we start importing
This means we now officially support putting the livestatus module
in the libexec/modules directory, which is nifty in case people
want to use our stuff but haven't got root access on the servers
where they actually need to run it.

As a sidenote, it's worth noting that each command module has had
code to add the module path themselves for quite a long time, but
only the 'node' command module (and everything the node command
module uses, which is more or less everything) has needed it, so
with this patch it becomes a teensy bit easier to write additional
command modules, and we can start removing quite a lot of code of
the cut 'n paste variety.

Signed-off-by: Andreas Ericsson <ae@op5.se>