daemon: Stop hardcoding the cache file path
Doing so means we have to keep track of it both in the module and
in the daemon, which is twice the work for no extra benefit.

While at it, fix it so we single-quote the cache- and status.log-
file names, just in case someone decides to put them in a path
with spaces.

Signed-off-by: Andreas Ericsson <ae@op5.se>