spec: Create test package for test files
As per the new convention that I just invented (actually, a centos
machine will return multiple results for 'yum search -- -test' so
there's some precedent), test scripts are available separately. For now, this
only applies to the mon test suite, but as it can do weird things like
disable active checks (although with a huge warning sign), we gain
nothing from including it in the default packages.

More test files should be moved/included in the test package as well, but
that's for later.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>