module: Block acknowledgement notifications originating from the net
Previously we used to check if we were running as the reaper thread
in order to block certain things (well, acknowledgements, really)
that originated from the network rather than the local node, but in
the anti-threading frenzy surrounding Nagios 4 that got sort of
lost.

This patch builds in a different mechanism which lets us distinguish
network-based events from locally generated ones and uses that to
block acknowledgement notifications that originated from another
node.

Reported-by: Mark Frost <mark.frost1@pepsico.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>