mon node ctrl: Warn and exit with error when we disregard all nodes
When 'mon node ctrl' is only passed nodenames which we're configured
not to connect to, we used to exit with a code indicating success.
Since that's not entirely accurate, and 'mon node ctrl' is used to
test commands on pollers, we really should do something better than
that. This patch makes it so that if only blocked nodes are given
we'll exit with return code 3 (conveniently, that's UNKNOWN to
Nagios) and print a message stating that we can't connect to that
node.

This sort of fixes issue #6452.

Signed-off-by: Andreas Ericsson <ae@op5.se>