import.c: Don't crash when importing downtimes
This bug was created by yours truly when attempting to port the hash
tables from the old, merlin-only API to the new, shared nagios API. With
the merlin ones, when there was a collision, the new entry would win.
With the nagios one, the opposite is true. This meant we would save the
wrong downtime entry when downtimes clashed - which they invariably do.

So, basically, I suck.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>