Fix multiple acknowledgement notifications, take two
With this patch we retain most of the old logic but introduce a
neat spot for handling network-command-received triggered events,
such as acknowledgement- and custom notifications.

Hopefully this will make it easier to fix future issues when
they crop up, while we're taking care of the current issue with
multiple acknowledgement notifications and avoiding multiple
recovery notifications.

Reported-by: Mark Frost <mark.frost1@pepsico.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>