mon test: Add valgrind capabilities
It's nifty to be able to run daemons through valgrind when testing,
and this patch lets us do just that. There's still no commandline
support to enable it, but one will quite easily find the valgrind
code and hopefully the trigger to turn it on with some casual
source-code browsing.

Signed-off-by: Andreas Ericsson <ae@op5.se>