mon node status: Aid more in peer debugging
It's pretty helpful to see the peer id of the systems we're
listing, but when it's in a different peer group than the
local node belong to, we never set the peer_id. In that case,
the only thing we can do is trust the self_assigned_peer_id
(which should be right almost always, but can be used to
spot discrepancies in user systems when people email the
status output in a bugreport.

To check for config errors, we also need to know how many
configured peers each system has, and how many each of those
peers the listed node thinks is active, so we print that too.

Signed-off-by: Andreas Ericsson <ae@op5.se>