module: Track statistics of notifications
This is sort of vital in order for us to be able to test that
certain notifications are only sent from certain systems. Since
notifications are at the top of our priority tree, it's really
important that we *do* test them.

We currently track notifications in a wide variety of categories.
For all the 9 different "reasons" (normal, custom, acknowledgement
and flapping/downtime start/stop/cancelled) we track host/service
notifications in both active and passive states. All in all 36
different types. We haven't bothered with "escalated or not" and
"recurring or not", since 36 scenarios is quite a bit nicer to
handle than the 144 it would be if we tried to catch all the ones
I could think of when writing this code.

We don't track how many notifications a particular contact gets,
and we don't distribute our contact stats to other nodes, but
this should be pretty useful even so.

Signed-off-by: Andreas Ericsson <ae@op5.se>