sql.c: Fix compiler warning about unused variable 'result'
The oracle driver returns TRUE (1) on success, and FALSE (0) on
failure. MySQL returns 0 on success and an error code on failure.
Since we never checked it before and the warning bugged the hell
out of me, I've shut the thing up with a cast to (void). Take
that, you silly thing...

Signed-off-by: Andreas Ericsson <ae@op5.se>