module: Simplify notification logic (take 2)
The last time we oversimplified it somewhat, leading to us sometimes
sending multiple notifications. With this patch, we retain most of
the old notification logic, but still implement the early return for
never allowing network-triggered events to cause us to send any
notifications what so ever, which lets us remove a lot of special
handling later in the function.

Signed-off-by: Andreas Ericsson <ae@op5.se>