module: Only allow 'normal' notifications for passive checks
The code to pass through notifications for passive checks from the
node that took in the command was a bit too lax and let notifications
for downtime, flapping and everything else that gets triggered on a
schedule slip through, causing multiple notifications for hosts and
services that were last checked using a passive check.

With this patch we should get rid of such things, and only passive
checks that trigger notifications should slip through.

Signed-off-by: Andreas Ericsson <ae@op5.se>