module: Only send human-cancelled downtime events across the network
When someone cancels downtime, we get a downtime stop event with
the attribute NEBATTR_DOWNTIME_CANCEL, which we must transfer
across the network. If the downtime expires normally, we must *not*
transfer it, since that may interfere with the scheduling queue in
case the clocks are off by roughly the network latency between the
systems and one node is attempting to delete downtime while the
command to do so from a different node is in the receive queue.

Signed-off-by: Andreas Ericsson <ae@op5.se>