mon test: Add ocount command to count compound config object types
This makes (manual) testing pretty trivial, and makes it possible
to create trivial configuration validators that ensure that all
objects that are supposed to be created actually are.

It can also count the number of downtimes and comments in status.log
files, which is nifty, since we want to do that when running tests in
order to verify that comments and such are added to retention data
files between reloads.

The driver behind the command is a simple function in the compound
config parser module, called "count_object_types()".

Signed-off-by: Andreas Ericsson <ae@op5.se>