mon test: Remake the 'mark' command to use files instead of db
Since we're moving away from databases to the point that not even
the tests use them anymore, we might as well make sure we log stuff
that happens in a format we can easily parse later.

With this change, we get files written in standard compound-config
format. Since we already have tools to analyze and parse such
files, we can easily extract all relevant info from them later and
use the info in various tests.

Signed-off-by: Andreas Ericsson <ae@op5.se>