Add 'mon test check'; The more-than-just-a-little clumsy test check
Since most people use active checks, we need to be able to test
those, even when they change state. The way we can do that quite
easily (and simulate a high I/O load while we're at it...) and
also change the states of selected objects based on the results
of their active checks is to use a check that reads state-files
when run and exits with the proper exit code and output that we
put in the file.

This patch adds such a command, and it's already proven pretty
useful.

Signed-off-by: Andreas Ericsson <ae@op5.se>