mon test: Avoid closing the daemon's stdout/stderr filedescriptors
Doing so causes a ton of read() errors that are very hard to
diagnose, as errno will point us to ENOTTY (which is perfectly
correct for the isatty() call when trying to determine if a
closed file is a tty or not).

Since c00692c3147e6396707f31370b3b5740036c7157 we've been carrying
this bug around, which was masked by the fact that monitor and merlin
tends to be quite quick on reconnecting. To ensure it doesn't return,
we put a big fat comment warning against closing the filedescriptor
so other clumsy retards can avoid repeating my mistake.

Now 'mon test dist' finishes with 100% success again.

Signed-off-by: Andreas Ericsson <ae@op5.se>